[breaking change] api response schema update.
when I tried to introduce the verify-email schema to the swagger docs,
I ran into the limitation of ApiResponse<T>, where I must provide a type
here.
I can work around the limitation by providing T=void, this seems to be a
smart solution, but open a can of worms: broken tests, ambigious
schema/type definitions.
I have decided to reintroduce the ApiMessageSchema, and refactor all the
types, and tests, and swagger docs.